requirements.txt
dj-database-url==0.4.1
Django==1.9.7
djangorestframework==3.3.3
gunicorn==19.6.0
itsdangerous==0.24
psycopg2==2.6.2
